# Basic speedups, json=>json + naive local cache (enable with USE_CACHE=1)