--- a/solar/solar/interfaces/db/redis_graph_db.py
+++ b/solar/solar/interfaces/db/redis_graph_db.py
@@ -1,4 +1,7 @@
-import json
+try:
+    import ujson as json
+except ImportError:
+    import json
 import redis
 import fakeredis