--- a/solar/solar/interfaces/orm.py
+++ b/solar/solar/interfaces/orm.py
@@ -53,6 +53,8 @@ def cache_me(store):
                 pass
             else:
                 val = f(obj, *args, **kwargs)
+                if not val:
+                    return
                 store[obj.id] = val
                 return val
         if USE_CACHE: