Before new backend arrives, I created this naive cache + I changed json to ujson.

Before `openstack.py deploy` ~ took 32 seconds, with those changes it's about ~14 seconds.