@pigmej 2015-10-22
More timings:
`openstack.py create_all` 
- with ~ 15 seconds
- without ~ 28 seconds
- without and without ujson ~ 33 seconds

stage
- with ~ 5,2 seconds
- without ~ 6,5 seconds
- without and without ujson ~ 7.2 seconds

process
- with ~ 2.2 seconds
- without ~ 3,4 seconds
- without and without ujson ~ 3,6


To use, 
`export USE_CACHE=1` and `sudo pip install ujson`


*warning* don't enable cache for celery and long running processes. Treat this as temporary solution.